// types.ts
export type UserData = {
    id: string;
    fullName: string;
    phone: string;
    email: string;
    gadgetDesc: string;
    jobType: string;
    fault: string;
    acessoriesIncluded: string;
    diagnosis: string;
    amountCharged: string;
    amountPaid: string;
    status: string;
    notes: string;
    CreatedAt: string;
  };